+++ src/main/kotlin/me/rafaelldi/aspire/database/DatabaseService.kt
@@ -1,30 +1,44 @@
 package me.rafaelldi.aspire.database
 
+import com.intellij.database.access.DatabaseCredentialsUi
 import com.intellij.database.dataSource.LocalDataSource
 import com.intellij.database.dataSource.LocalDataSourceManager
 import com.intellij.database.util.DbImplUtil
+import com.intellij.database.util.LoaderContext
+import com.intellij.database.util.performAutoIntrospection
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
+import com.intellij.openapi.diagnostic.debug
+import com.intellij.openapi.diagnostic.logger
 import com.intellij.openapi.project.Project
+import com.intellij.platform.ide.progress.withBackgroundProgress
 import com.intellij.util.application
+import com.jetbrains.rider.plugins.appender.database.dialog.steps.shared.services.connection.ConnectionManager
+import com.jetbrains.rider.plugins.appender.database.dialog.steps.shared.services.connection.TestConnectionExecutionResult
 import com.jetbrains.rider.plugins.appender.database.jdbcToConnectionString.converters.ConnectionStringToJdbcUrlConverter
 import com.jetbrains.rider.plugins.appender.database.jdbcToConnectionString.dataProviders.MySqlClientDataProvider
 import com.jetbrains.rider.plugins.appender.database.jdbcToConnectionString.dataProviders.NpgsqlDataProvider
 import com.jetbrains.rider.plugins.appender.database.jdbcToConnectionString.dataProviders.OracleClientDataProvider
 import com.jetbrains.rider.plugins.appender.database.jdbcToConnectionString.dataProviders.SqlClientDataProvider
 import com.jetbrains.rider.plugins.appender.database.jdbcToConnectionString.factories.ConnectionStringsFactory
+import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.channels.BufferOverflow
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.launch
 import me.rafaelldi.aspire.settings.AspireSettings
+import kotlin.time.Duration.Companion.milliseconds
 
 @Service(Service.Level.PROJECT)
 class DatabaseService(private val project: Project, scope: CoroutineScope) {
     companion object {
         fun getInstance(project: Project) = project.service<DatabaseService>()
 
-        private const val REDIS_CONNECTION_STRING_PATTERN = "(?<host>\\w*):(?<port>\\d*)(,user=(?<user>\\w*))?(,password=(?<password>\\w*))?"
+        private val LOG = logger<DatabaseService>()
+
+        private const val REDIS_CONNECTION_STRING_PATTERN =
+            "(?<host>\\w*):(?<port>\\d*)(,user=(?<user>\\w*))?(,password=(?<password>\\w*))?"
         private val REDIS_REGEX = Regex(REDIS_CONNECTION_STRING_PATTERN)
     }
 
@@ -32,6 +46,8 @@ class DatabaseService(private val project: Project, scope: CoroutineScope) {
         DatabaseResourceType.MSSQL,
         DatabaseResourceType.MONGO
     )
+
+    private val connectionManager = ConnectionManager(project)
 
     private val connectionStrings = mutableSetOf<DatabaseResourceConnectionString>()
     private val databaseResources = mutableSetOf<DatabaseResource>()
@@ -43,9 +59,17 @@ class DatabaseService(private val project: Project, scope: CoroutineScope) {
         extraBufferCapacity = 100
     )
 
+    private val createdDataSources = MutableSharedFlow<LocalDataSource>(
+        onBufferOverflow = BufferOverflow.DROP_OLDEST,
+        extraBufferCapacity = 100
+    )
+
     init {
         scope.launch {
             commands.collect { handleCommand(it) }
+        }
+        scope.launch {
+            createdDataSources.collect { handleDataSource(it) }
         }
     }
 
@@ -153,8 +177,10 @@ class DatabaseService(private val project: Project, scope: CoroutineScope) {
 
         val dataSource = LocalDataSource.fromDriver(driver, url, true).apply {
             name = connectionString.name
+            isAutoSynchronize = true
         }
         val dataSourceManager = LocalDataSourceManager.getInstance(project)
+
         resource.lifetime.bracketIfAlive({
             synchronized(lock) {
                 createdConnections.add(connectionString.connectionString)
@@ -170,6 +196,8 @@ class DatabaseService(private val project: Project, scope: CoroutineScope) {
                 dataSourceManager.removeDataSource(dataSource)
             }
         })
+
+        createdDataSources.tryEmit(dataSource)
     }
 
     private fun convertRedisConnectionString(connectionString: String): String? {
@@ -195,6 +223,45 @@ class DatabaseService(private val project: Project, scope: CoroutineScope) {
 
     private fun handleRemoveDatabaseResourceCommand(command: RemoveDatabaseResourceCommand) {
         databaseResources.remove(command.resource)
+    }
+
+    private suspend fun handleDataSource(dataSource: LocalDataSource) =
+        withBackgroundProgress(project, "Connecting to the database...") {
+            val isConnectionSuccessful = waitForConnection(dataSource)
+            if (!isConnectionSuccessful) {
+                LOG.warn("Unable to connect to database")
+                return@withBackgroundProgress
+            }
+
+            try {
+                performAutoIntrospection(LoaderContext.selectGeneralTask(project, dataSource), true)
+            } catch (ce: CancellationException) {
+                LOG.trace("Introspection is canceled")
+            } catch (e: Exception) {
+                LOG.warn("Unable to perform auto introspection", e)
+            }
+        }
+
+    private suspend fun waitForConnection(dataSource: LocalDataSource): Boolean {
+        val credentials = DatabaseCredentialsUi.newUIInstance()
+
+        for (i in 1..<5) {
+            when (val connectionResult = connectionManager.testConnection(dataSource, credentials)) {
+                TestConnectionExecutionResult.Cancelled -> {
+                    LOG.debug("Connection cancelled")
+                    return false
+                }
+
+                is TestConnectionExecutionResult.Failure -> {
+                    LOG.debug { "Unable to connect to database, ${connectionResult.result.summary}" }
+                    delay(300.milliseconds)
+                }
+
+                is TestConnectionExecutionResult.Success -> return true
+            }
+        }
+
+        return false
     }
 
     private interface DatabaseResourceCommand