+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Unit testing support
 - Target framework, arguments, working directory and podman runtime parameters to the run configuration
+- Automatic introspection for the created database
 
 ## [1.3.1] - 2024-07-17